terhapus:        tests/Feature/Action/Content/ArticleActionTest.php 	terhapus:        tests/Feature/Action/Content/CategoryActionTest.php 	terhapus:        tests/Feature/Action/Content/CommentActionTest.php 	terhapus:        tests/Feature/Action/Content/PageActionTest.php 	terhapus:        tests/Feature/Action/Sid/PendudukActionTest.php 	terubah:         tests/Feature/CreateTeamTest.php 	terubah:         tests/Feature/DeleteTeamTest.php 	terubah:         tests/Feature/LeaveTeamTest.php 	terubah:         tests/Feature/RemoveTeamMemberTest.php 	terubah:         tests/Feature/UpdateTeamMemberRoleTest.php 	terubah:         tests/Feature/UpdateTeamNameTest.php

+++ tests/Feature/LeaveTeamTest.php
@@ -4,6 +4,7 @@ namespace Tests\Feature;
 
 use App\Models\User;
 use Illuminate\Foundation\Testing\RefreshDatabase;
+use Laravel\Jetstream\Features;
 use Tests\TestCase;
 
 class LeaveTeamTest extends TestCase
@@ -12,6 +13,12 @@ class LeaveTeamTest extends TestCase
 
     public function testUsersCanLeaveTeams(): void
     {
+        if (!Features::hasTeamFeatures()) {
+            $this->markTestSkipped('Team feature is not enabled.');
+
+            return;
+        }
+
         $user = User::factory()->withPersonalTeam()->create();
 
         $user->currentTeam->users()->attach(
@@ -27,6 +34,12 @@ class LeaveTeamTest extends TestCase
 
     public function testTeamOwnersCantLeaveTheirOwnTeam(): void
     {
+        if (!Features::hasTeamFeatures()) {
+            $this->markTestSkipped('Team feature is not enabled.');
+
+            return;
+        }
+
         $this->actingAs($user = User::factory()->withPersonalTeam()->create());
 
         $response = $this->delete('/teams/'.$user->currentTeam->id.'/members/'.$user->id);
